% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{ab_hr}
\alias{ab_hr}
\title{Calculates at bats per home run}
\usage{
ab_hr(ab, hr)
}
\arguments{
\item{ab}{Number of at bats}

\item{hr}{Number of home runs}
}
\value{
ab_hr
}
\description{
Takes number of at bats and divides by number of home runs
}
\examples{
ab_hr(400, 25)
}
