% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{ba}
\alias{ba}
\title{Calculates batting average}
\usage{
ba(h, ab)
}
\arguments{
\item{h}{Number of hits}

\item{ab}{Number of at bats}
}
\value{
ba
}
\description{
Takes number of hits and divides by at bats. 1.000 (read ``one-thousand" is perfect)
}
\examples{
ba(200, 525)
}
