% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{baa}
\alias{baa}
\title{Calculates batting average against}
\usage{
baa(h, bfp, bb, hbp, sh, sf, ci)
}
\arguments{
\item{h}{Number of hits allowed}

\item{bfp}{Number of batters facing pitcher}

\item{bb}{Number of bases on balls}

\item{hbp}{Number of hit batters}

\item{sh}{Number of sacrifice hits}

\item{sf}{Number of sacrifice flies}

\item{ci}{Number of catcher's interference}
}
\value{
baa
}
\description{
Computes pitcher's ability to prevent hits, based on h, bfp, bb, hbp, sf, sh, and ci (catcher's interference)
}
\examples{
baa(105, 250, 50, 15, 10, 5, 1)
}
