% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babip.R
\name{babip}
\alias{babip}
\title{Batting Average on Balls in Play FUNCTION}
\usage{
babip(x)
}
\arguments{
\item{x}{Dataset containing batting statistics or pitching statistics}
}
\description{
This function measures the frequency of batter's ball in play that go for hits
}
\examples{
babip()
}
\keyword{babip}
