% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{bb9}
\alias{bb9}
\title{Calculates bases on balls per nine innings pitched (W/9)}
\usage{
bb9(bb, ip)
}
\arguments{
\item{bb}{Number of bases on balls}

\item{ip}{Number of innings pitched}
}
\value{
bb9
}
\description{
Computes bases on balls (walks) per nine innings pitched
}
\examples{
bb9(35, 210)
}
