% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{BsR}
\alias{BsR}
\title{Calculates the base runs estimator}
\usage{
BsR(h, bb, hr, tb, ab)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{hr}{Number of home runs}

\item{tb}{Number of total bases (one for 1B, two for 2B, three for 3B, and four for HR)}

\item{ab}{Number of at bats}
}
\value{
BsR
}
\description{
Takes the number of hits, bases on balls, home runs, total bases, and at bats to compute the base runs estimator, which is similar to runs created
}
\examples{
BsR(135, 22, 12, 155, 330)
}
