% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct}
\alias{ct}
\title{CONTACT RATE FUNCTION}
\usage{
ct(x)
}
\arguments{
\item{x}{Dataset containing batting statistics}
}
\description{
This function measures the ratio of times the batter makes contact with the ball
}
\examples{
ct()
}
\keyword{ct}
