% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{DEFENSE-INDEPENDENT COMPONENT ERA FUNCTION}
\usage{
dice(x)
}
\arguments{
\item{x}{Dataset containing pitching statistics}
}
\description{
This function estimates a pitcher's ERA based on defense-independent stats
}
\examples{
dice()
}
\keyword{dice}
