% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{EqA}
\alias{EqA}
\title{Calculates equivalent average}
\usage{
EqA(h, tb, bb, hbp, sb, sh, sf, ab, cs)
}
\arguments{
\item{h}{Number of hits}

\item{tb}{Number of total bases (one for 1B, two for 2B, three for 3B, and four for HR)}

\item{bb}{Number of bases on balls}

\item{hbp}{Number of hits by pitch}

\item{sb}{Number of stolen bases}

\item{sh}{Number of sacrifice hits (typically bunts)}

\item{sf}{Number of sacrifice flies}

\item{ab}{Number of at bats}

\item{cs}{Number of caught stealing}
}
\value{
EqA
}
\description{
Takes the number of hits, total bases, bases on balls, hits by pitch, stolen bases, sacrifice hits, sacrifice flies, at bats, and caught stealing to compute the base runs, which is a player's batting average absent park and league effects
}
\examples{
EqA(135, 155, 22, 3, 15, 4, 2, 365, 1)
}
