% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{era}
\alias{era}
\title{Calculates earned run average}
\usage{
era(er, ip)
}
\arguments{
\item{er}{Number of runs that did not occur as a result of errors or passed balls}

\item{ip}{Number of innings pitched}
}
\value{
era
}
\description{
Computes a pitcher's earned run average (compare with "erc" which is the component earned run average, "aera" which is a pitcher's adjusted earned run average, or "dice" which is the defense-independent component earned run average)
}
\examples{
era(150, 400)
}
