% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erc.R
\name{erc}
\alias{erc}
\title{COMPONENT ERA FUNCTION}
\usage{
erc(x)
}
\arguments{
\item{x}{Dataset containing pitching statistics}
}
\description{
This function estimates a pitcher's ERA
}
\examples{
erc()
}
\keyword{erc}
