% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{fip}
\alias{fip}
\title{Calculates fielding independent pitching}
\usage{
fip(hr, bb, k, ip)
}
\arguments{
\item{hr}{Number of home runs}

\item{bb}{Number of bases on balls}

\item{k}{Number of strikeouts}

\item{ip}{Number of innings pitched}
}
\value{
fip
}
\description{
Computes pitching performance statistic similar to ERA, but based on factors within the pitcher's control (compare with "dice" which is the defense-independent component earned run average)
}
\examples{
fip(65, 50, 100, 175)
}
