% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielding.R
\name{fp}
\alias{fp}
\title{Calculates fielding percentage}
\usage{
fp(p, a, e)
}
\arguments{
\item{p}{Number of putouts}

\item{a}{Number of assists}

\item{e}{Number of errors}
}
\value{
fp
}
\description{
Computes the fielding percentage (aka, fielding average), which reflects the percentage of proper ball handling
}
\examples{
fp(13, 4, 2)
}
