% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{go_ao}
\alias{go_ao}
\title{Calculates ground outs-fly outs ratio (GO/AO)}
\usage{
go_ao(go, ao)
}
\arguments{
\item{go}{Number of ground ball outs}

\item{ao}{Number of fly ball outs}
}
\value{
go_ao
}
\description{
Takes the number of ground ball outs and divides by number of fly ball outs to compute the GO/AO ratio
}
\examples{
go_ao(150, 88)
}
