% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{gpa}
\alias{gpa}
\title{GROSS PRODUCTION AVERAGE FUNCTION}
\usage{
gpa(x)
}
\arguments{
\item{x}{Dataset containing batting statistics}
}
\description{
This function measures a batter's ability to score runs
}
\examples{
gpa()
}
\keyword{gpa}
