% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{h9}
\alias{h9}
\title{Calculates hits per nine innings pitched (H/9IP)}
\usage{
h9(h, ip)
}
\arguments{
\item{h}{Number of hits allowed}

\item{ip}{Number of innings pitched}
}
\value{
h9
}
\description{
Computes hits per nine innings pitched
}
\examples{
h9(150, 175)
}
