% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.R
\name{iso}
\alias{iso}
\title{ISOLATED POWER FUNCTION}
\usage{
iso(x)
}
\arguments{
\item{x}{Dataset containing batting statistics}
}
\description{
This function measures a batter's ability to hit for extra bases
}
\examples{
iso()
}
\keyword{iso}
