% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{k9}
\alias{k9}
\title{Calculates strikeouts per nine innings pitched (K/9)}
\usage{
k9(k, ip)
}
\arguments{
\item{k}{Number of strikeouts}

\item{ip}{Number of innings pitched}
}
\value{
k9
}
\description{
Computes strikeouts per nine innings pitched
}
\examples{
k9(105, 175)
}
