% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{k_bb}
\alias{k_bb}
\title{Calculates strikeout to walk ratio (pitching)}
\usage{
k_bb(k, bb)
}
\arguments{
\item{k}{Number of strikeouts}

\item{bb}{Number of bases on balls}
}
\value{
k_bb
}
\description{
Computes strikeouts to walk ratio, based on number of strikeouts and number of walks  (for batting version, see "bb_k")
}
\examples{
k_bb(105, 40)
}
