% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{obp}
\alias{obp}
\title{Calculates on-base percentage}
\usage{
obp(h, bb, hbp, ab, sf)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{hbp}{Number of hits by pitch}

\item{ab}{Number of at bats}

\item{sf}{Number of sacrifice flies}
}
\value{
obp
}
\description{
Computes the on-base percentage based on number of hits, bases on balls, hits by pitch, at bats, and sacrifice flies
}
\examples{
obp(150, 40, 2, 400, 5)
}
