% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{ops}
\alias{ops}
\title{Calculates on-base plus slugging}
\usage{
ops(h, bb, hbp, ab, sf, b1, b2, b3, hr)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{hbp}{Number of hits by pitch}

\item{ab}{Number of at bats}

\item{sf}{Number of sacrifice flies}

\item{b1}{Number of singles}

\item{b2}{Number of doubles}

\item{b3}{Number of triples}

\item{hr}{Number of home runs}
}
\value{
ops
}
\description{
Computes the on-base percentage plus slugging average (OPS) based on number of hits, bases on balls, hits by pitch, at bats, sacrifice flies, and total weighted bases (represented individually, as in SLG and GPA calculations)
}
\examples{
ops(200, 18, 4, 401, 4, 50, 20, 3, 13)
}
