% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{pa_so}
\alias{pa_so}
\title{Calculates plate appearances per strikeout (PA/SO)}
\usage{
pa_so(pa, so)
}
\arguments{
\item{pa}{Number of plate appearances}

\item{so}{Number of strikeouts}
}
\value{
pa_so
}
\description{
Computes the number of times a hitter strikes out in relation to their plate appearances
}
\examples{
pa_so(450, 120)
}
