% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielding.R
\name{pafa}
\alias{pafa}
\title{Calculates park factor}
\usage{
pafa(home_rs, home_ra, home_r, road_rs, road_ra, road_r)
}
\arguments{
\item{home_rs}{Number of pitcher's team runs scored at home park}

\item{home_ra}{Number of pitcher's team runs allowed at home park}

\item{home_r}{Total number of runs scored at home park}

\item{road_rs}{Number of pitcher's team runs scored at away park}

\item{road_ra}{Number of pitcher's team runs allowed at away park}

\item{road_r}{Total number of runs scored at away park}
}
\value{
pafa
}
\description{
Computes the runs a team scores at home versus away (it is often used in other metrics, e.g., adjusted era (ERA+) for pitchers; see "aera")
}
\examples{
pafa(5, 6, 11, 4, 8, 12) # for a pair of games (one home, one away)
}
