% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptb.R
\name{ptb}
\alias{ptb}
\title{PITCHER'S TOTAL BASES FUNCTION}
\usage{
ptb(x)
}
\arguments{
\item{x}{Dataset containing pitching statistics}
}
\description{
This function measures total baserunners allowed by pitcher
}
\examples{
ptb()
}
\keyword{ptb}
