% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{ra}
\alias{ra}
\title{Calculates run average}
\usage{
ra(r, ip)
}
\arguments{
\item{r}{Number of runs allowed}

\item{ip}{Number of innings pitched}
}
\value{
ra
}
\description{
Computes pitcher's run average based on number of runs allowed and innings pitched
}
\examples{
ra(75, 400)
}
