% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc}
\alias{rc}
\title{RUNS CREATED FUNCTION}
\usage{
rc(x)
}
\arguments{
\item{x}{Dataset containing batting statistics}
}
\description{
This function measures how many runs a batter has contributed
}
\examples{
rc()
}
\keyword{rc}
