% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{rc2}
\alias{rc2}
\title{Calculates runs created accounting for stolen bases}
\usage{
rc2(h, bb, tb, ab, cs, sb)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{tb}{Number of total bases (one for 1B, two for 2B, three for 3B, and four for HR)}

\item{ab}{Number of at bats}

\item{cs}{Number of stolen bases caught}

\item{sb}{Number of stolen bases}
}
\value{
rc2
}
\description{
Computes the estimated runs a hitter creates or contributes, accounting for base stealing (see also "rc" for the basic iteration and "rc3" for the technical iteration of the rc statistic)
}
\examples{
rc2(150, 35, 165, 400, 7, 9)
}
