% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{rc3}
\alias{rc3}
\title{Calculates runs created accounting for all offensive indicators}
\usage{
rc3(h, bb, ibb, tb, ab, cs, sb, hbp, gidp, sh, sf)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{ibb}{Number of intentional bases on balls}

\item{tb}{Number of total bases (one for 1B, two for 2B, three for 3B, and four for HR)}

\item{ab}{Number of at bats}

\item{cs}{Number of stolen bases caught}

\item{sb}{Number of stolen bases}

\item{hbp}{Number of hits by pitch}

\item{gidp}{Number of grounded into double play}

\item{sh}{Number of sacrifice hits}

\item{sf}{Number of sacrifice flies}
}
\value{
rc3
}
\description{
Computes the technical iteration of estimated runs a hitter creates or contributes accounting for virtually all offensive indicators (see also "rc" for the basic iteration and "rc2" for the 'stolen base' iteration of the rc statistic)
}
\examples{
rc3(150, 35, 3, 165, 400, 7, 9, 5, 1, 6, 2)
}
