% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielding.R
\name{rfa}
\alias{rfa}
\title{Calculates range factor}
\usage{
rfa(p, a, ip)
}
\arguments{
\item{p}{Number of putouts}

\item{a}{Number of assists}

\item{ip}{Number of innings played in a defensive position}
}
\value{
rfa
}
\description{
Computes the amount of the field covered by a player
}
\examples{
rfa(20, 5, 450)
}
