% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{risp}
\alias{risp}
\title{Calculates batting average with runners in scoring position}
\usage{
risp(hrisp, abrisp)
}
\arguments{
\item{hrisp}{Number of hits with runners in scoring position (on either 2nd or 3rd base)}

\item{abrisp}{Number of at bats with runners in scoring position (on either 2nd or 3rd base)}
}
\value{
risp
}
\description{
Computes batting average accounting for runners in scoring position
}
\examples{
risp(35, 120)
}
