% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{rp}
\alias{rp}
\title{Calculates runs produced}
\usage{
rp(r, rbi, hr)
}
\arguments{
\item{r}{Number of runs}

\item{rbi}{Number of runs batted in}

\item{hr}{Number of home runs}
}
\value{
rp
}
\description{
Computes the number of runs contributed by a hitter, based on runs, runs batted in, and home runs
}
\examples{
rp(70, 41, 22)
}
