% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baserunning.R
\name{sba}
\alias{sba}
\title{Calculates stolen base attempts}
\usage{
sba(sb, cs)
}
\arguments{
\item{sb}{Number of stolen bases}

\item{cs}{Number of caught stealing}
}
\value{
sba
}
\description{
Computes total attempts to steal a base, by adding sb and cs
}
\examples{
sba(20, 4)
}
