% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baserunning.R
\name{sbsr}
\alias{sbsr}
\title{Calculates stolen base success rate}
\usage{
sbsr(sb, cs)
}
\arguments{
\item{sb}{Number of stolen bases}

\item{cs}{Number of caught stealing}
}
\value{
sbsr
}
\description{
Computes percentage of bases successfully stolen
}
\examples{
sbsr(20, 4)
}
