% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{slg}
\alias{slg}
\title{Calculates slugging percentage}
\usage{
slg(b1, b2, b3, hr, ab)
}
\arguments{
\item{b1}{Number of singles}

\item{b2}{Number of doubles}

\item{b3}{Number of triples}

\item{hr}{Number of home runs}

\item{ab}{Number of at bats}
}
\value{
slg
}
\description{
Computes the slugging percentage (SLG), based on the weighted number of singles, doubles, triples, home runs, and at bats
}
\examples{
slg(100, 40, 3, 7, 350)
}
