% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{ta}
\alias{ta}
\title{Calculates total average}
\usage{
ta(tb, hbp, bb, sb, ab, h, cs, gidp)
}
\arguments{
\item{tb}{Number of total bases (one for 1B, two for 2B, three for 3B, and four for HR)}

\item{hbp}{Number of hits by pitch}

\item{bb}{Number of bases on balls}

\item{sb}{Number of stolen bases}

\item{ab}{Number of at bats}

\item{h}{Number of hits}

\item{cs}{Number of caught stealing}

\item{gidp}{Number of grounded into double play}
}
\value{
ta
}
\description{
Computes overall offensive contribution of a single player
}
\examples{
ta(125, 11, 40, 10, 400, 105, 2, 6)
}
