% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielding.R
\name{tc}
\alias{tc}
\title{Calculates total chances}
\usage{
tc(p, a, e)
}
\arguments{
\item{p}{Number of putouts}

\item{a}{Number of assists}

\item{e}{Number of errors}
}
\value{
tc
}
\description{
Computes the opportunities for defensive ball handling
}
\examples{
tc(11, 5, 5)
}
