% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{tob}
\alias{tob}
\title{Calculates times on base}
\usage{
tob(h, hbp, bb)
}
\arguments{
\item{h}{Number of hits}

\item{hbp}{Number of hits by pitch}

\item{bb}{Number of bases on balls}
}
\value{
tob
}
\description{
Computes total times a player reaches a base by adding h, hbp, and bb
}
\examples{
tob(234, 6, 24)
}
