% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{vorp}
\alias{vorp}
\title{Calculates value over replacement player (pitching)}
\usage{
vorp(ip, lr, lg, r)
}
\arguments{
\item{ip}{Number of innings pitched}

\item{lr}{Number of league runs}

\item{lg}{Number of league games played}

\item{r}{Number of runs}
}
\value{
vorp
}
\description{
Computes a pitcher's marginal utility
}
\examples{
vorp(400, 98, 20, 110)
}
