% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{whip}
\alias{whip}
\title{Calculates walks plus hits per innings pitched (WHIP)}
\usage{
whip(bb, h, ip)
}
\arguments{
\item{bb}{Number of bases on balls}

\item{h}{Number of hits allowed}

\item{ip}{Number of innings pitched}
}
\value{
whip
}
\description{
Computes walks plus hits per innings pitched, which reflects the number of baserunners allowed by a pitcher over a given period
}
\examples{
whip(50, 110, 425)
}
