% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{wr}
\alias{wr}
\title{Calculates whiff rate}
\usage{
wr(sw, tp)
}
\arguments{
\item{sw}{Number of swings and misses}

\item{tp}{Total pitches thrown}
}
\value{
wr
}
\description{
Computes pitcher's ability to get a batter to swing and miss pitches over any period of time (e.g., in a single game, single season, career, etc.)
}
\examples{
wr(300, 750)
}
