% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielding.R
\name{wra}
\alias{wra}
\title{Calculates win ratio}
\usage{
wra(rs, ra)
}
\arguments{
\item{rs}{Number of runs scored}

\item{ra}{Number of runs allowed}
}
\value{
wra
}
\description{
Computes a team's win ratio, which is used in the so-called "Pythagorean expectation"
}
\examples{
wra(400, 301)
}
