% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{xbh}
\alias{xbh}
\title{Calculates extra base hits}
\usage{
xbh(b2, b3, hr)
}
\arguments{
\item{b2}{Number of doubles}

\item{b3}{Number of triples}

\item{hr}{Number of home runs}
}
\value{
xbh
}
\description{
Computes total hits by a player greater than singles (1B) by adding 2B, 3B, and hr
}
\examples{
xbh(20, 18, 4)
}
