% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{babip}
\alias{babip}
\title{Calculates batting average on balls in play}
\usage{
babip(h, hr, ab, k, sf)
}
\arguments{
\item{h}{Number of hits}

\item{hr}{Number of home runs}

\item{ab}{Number of at bats}

\item{k}{Number of strikeouts}

\item{sf}{Number of sacrifice flies}
}
\value{
babip
}
\description{
Generates the frequency a batter reaches a base after putting the ball in play (normal around .300)
}
\examples{
babip(200, 25, 525, 55, 6)
}
