% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{dice}
\alias{dice}
\title{Calculates defense-independent component earned run average}
\usage{
dice(bb, hbp, hr, k, ip)
}
\arguments{
\item{bb}{Number of bases on balls}

\item{hbp}{Number of hit batters}

\item{hr}{Number of home runs}

\item{k}{Number of strikeouts}

\item{ip}{Number of innings pitched}
}
\value{
dice
}
\description{
Computes earned run average from hits and walks (compare with "era" which is the traditional formula for earned run average, "aera" which is a pitcher's adjusted earned run average, or "erc" which is the component earned run average)
}
\examples{
dice(45, 10, 60, 130, 400)
}
