% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitching.R
\name{erc}
\alias{erc}
\title{Calculates component earned run average}
\usage{
erc(h, bb, hbp, hr, ibb, bfp, ip)
}
\arguments{
\item{h}{Number of hits allowed}

\item{bb}{Number of bases on balls}

\item{hbp}{Number of hit batters}

\item{hr}{Number of home runs}

\item{ibb}{Number of intentional bases on balls}

\item{bfp}{Number of batters faced by pitcher}

\item{ip}{Number of innings pitched}
}
\value{
erc
}
\description{
Computes earned run average from hits and walks (compare with "era" which is the traditional formula for earned run average, "aera" which is a pitcher's adjusted earned run average, or "dice" which is the defense-independent component earned run average)
}
\examples{
erc(110, 45, 10, 70, 5, 400, 215)
}
