% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{gpa}
\alias{gpa}
\title{Calculates gross production average}
\usage{
gpa(h, bb, hbp, ab, sf, b1, b2, b3, hr)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{hbp}{Number of hits by pitch}

\item{ab}{Number of at bats}

\item{sf}{Number of sacrifice flies}

\item{b1}{Number of singles}

\item{b2}{Number of doubles}

\item{b3}{Number of triples}

\item{hr}{Number of home runs}
}
\value{
gpa
}
\description{
Computes the gross production average, which is 1.8 times on-base percentage (OBP) plus slugging percentage (SLG), divided by four
}
\examples{
gpa(150, 40, 2, 400, 5, 100, 40, 3, 7)
}
