% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{iso}
\alias{iso}
\title{Calculates isolated power}
\usage{
iso(b1, b2, b3, hr, ab, h)
}
\arguments{
\item{b1}{Number of singles}

\item{b2}{Number of doubles}

\item{b3}{Number of triples}

\item{hr}{Number of home runs}

\item{ab}{Number of at bats}

\item{h}{Number of hits}
}
\value{
iso
}
\description{
Computes isolated power, which is a player's ability to obtain extra bases from a hit. The statistic subtracts a hitter's batting average from the slugging percentage, with the maximum ISO being 3.000.
}
\examples{
iso(100, 40, 3, 7, 350, 150)
}
