% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batting.R
\name{rc}
\alias{rc}
\title{Calculates runs created}
\usage{
rc(h, bb, tb, ab)
}
\arguments{
\item{h}{Number of hits}

\item{bb}{Number of bases on balls}

\item{tb}{Number of total bases (one for 1B, two for 2B, three for 3B, and four for HR)}

\item{ab}{Number of at bats}
}
\value{
rc
}
\description{
Computes the basic version of the estimated runs a hitter creates or contributes (see also "rc2" for the 'stolen base' iteration and "rc3" for the technical iteration of the rc statistic)
}
\examples{
rc(150, 35, 165, 400)
}
